requirements.txt
asgiref==3.6.0
Django==4.1.7
django-dotenv==1.4.2
django-storages==1.13.2
psycopg2==2.9.5
sqlparse==0.4.3
